+++ client/assets/js/components/forms/RemediationTaskTemplateForm.js
@@ -90,6 +90,10 @@ const RemediationTaskTemplateForm = {
             <option value="pif">PIF</option>
             <option value="bond">Bond</option>
             <option value="vlan">VLAN</option>
+            <option value="vmss">VM Snapshot Schedule</option>
+            <option value="host_patch">Host Patch</option>
+            <option value="certificate">Certificate</option>
+            <option value="pvs_proxy">PVS Proxy</option>
             <option value="task">Task</option>
             <option value="alert">Alert</option>
           </select>